refactor(meal-form): reorganize meal records store structure
This commit restructures the import paths for the meal records store files in the meal form feature. It moves `mealRecords.ts` into a new `stores` directory under `meal-form`, ensuring better encapsulation of related functionality. The imports in various components and utilities have been updated accordingly to reflect this change.

This refactoring aims to improve code maintainability by grouping related functionalities together, making it easier for developers to navigate the project structure. No changes were made to the underlying logic or data handling, ensuring that existing functionality remains intact.
