+++ src/features/meal-form/components/MealFormAccordionItem.tsx
@@ -3,8 +3,9 @@ import { useDisclosure } from '@mantine/hooks'
 import { useEffect } from 'react'
 
 import { DialogSaved } from '../../../components/DialogSaved'
-import { useCurrentDateStore, useMealRecordsStore } from '../../../stores'
+import { useCurrentDateStore } from '../../../stores'
 import { createStringFromDate } from '../../../utils'
+import { useMealRecordsStore } from '../stores'
 import { FormsType } from '../types'
 import { createFoodInitialValues, saveMealRecord } from '../utils'
 import { MealFormFields } from './MealFormFields'
