+++ src/stores/index.ts
@@ -1,5 +1,4 @@
 export * from './currentDate'
 export * from './dailyGoal'
-export * from './mealRecords'
 export * from './nutritionNumbers'
 export * from './weeklyMealRecords'
